refactor: recipes page controller
